+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/Rules/Ledger.hs
@@ -61,6 +61,7 @@ instance
   , Environment (EraRule "DELEGS" era) ~ DelegsEnv era
   , State (EraRule "DELEGS" era) ~ CertState era
   , Signal (EraRule "DELEGS" era) ~ Seq (TxCert era)
+  , ProtVerAtMost era 8
   ) =>
   STS (BabbageLEDGER era)
   where