+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Utxow.hs
@@ -1,0 +1,88 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DerivingVia #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE UndecidableInstances #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+module Cardano.Ledger.Conway.Rules.Utxow (
+  ConwayUTXOW,
+)
+where
+
+import Cardano.Crypto.DSIGN.Class (Signable)
+import Cardano.Crypto.Hash.Class (Hash)
+import Cardano.Ledger.Alonzo.Rules (
+  AlonzoUtxowEvent (WrappedShelleyEraEvent),
+ )
+import Cardano.Ledger.Alonzo.Rules as Alonzo (AlonzoUtxoEvent)
+import Cardano.Ledger.Alonzo.Scripts (AlonzoScript)
+import Cardano.Ledger.Alonzo.Tx (AlonzoEraTx)
+import Cardano.Ledger.Alonzo.TxInfo (ExtendedUTxO (..))
+import Cardano.Ledger.Alonzo.UTxO (AlonzoScriptsNeeded)
+import Cardano.Ledger.Babbage.Rules (
+  BabbageUTXO,
+  BabbageUtxoPredFailure,
+  BabbageUtxowPredFailure (..),
+ )
+import Cardano.Ledger.BaseTypes (ShelleyBase)
+import Cardano.Ledger.Conway.Core
+import Cardano.Ledger.Conway.Era (ConwayUTXOW)
+import Cardano.Ledger.Crypto (DSIGN, HASH)
+import Cardano.Ledger.Shelley.LedgerState (UTxOState (..))
+import Cardano.Ledger.Shelley.Rules (
+  ShelleyUtxowEvent (UtxoEvent),
+  UtxoEnv (..),
+ )
+import Cardano.Ledger.UTxO (EraUTxO (..))
+import Control.State.Transition.Extended (
+  Embed (..),
+  STS (..),
+ )
+
+instance
+  forall era.
+  ( ExtendedUTxO era
+  , AlonzoEraTx era
+  , EraUTxO era
+  , ScriptsNeeded era ~ AlonzoScriptsNeeded era
+  , ConwayEraTxBody era
+  , Signable (DSIGN (EraCrypto era)) (Hash (HASH (EraCrypto era)) EraIndependentTxBody)
+  , Script era ~ AlonzoScript era
+  , -- Allow UTXOW to call UTXO
+    Embed (EraRule "UTXO" era) (ConwayUTXOW era)
+  , Environment (EraRule "UTXO" era) ~ UtxoEnv era
+  , State (EraRule "UTXO" era) ~ UTxOState era
+  , Signal (EraRule "UTXO" era) ~ Tx era
+  , Eq (PredicateFailure (EraRule "UTXOS" era))
+  , Show (PredicateFailure (EraRule "UTXOS" era))
+  ) =>
+  STS (ConwayUTXOW era)
+  where
+  type State (ConwayUTXOW era) = UTxOState era
+  type Signal (ConwayUTXOW era) = Tx era
+  type Environment (ConwayUTXOW era) = UtxoEnv era
+  type BaseM (ConwayUTXOW era) = ShelleyBase
+  type PredicateFailure (ConwayUTXOW era) = BabbageUtxowPredFailure era
+  type Event (ConwayUTXOW era) = AlonzoUtxowEvent era
+  transitionRules = []
+  initialRules = []
+
+instance
+  ( Era era
+  , STS (BabbageUTXO era)
+  , PredicateFailure (EraRule "UTXO" era) ~ BabbageUtxoPredFailure era
+  , Event (EraRule "UTXO" era) ~ AlonzoUtxoEvent era
+  , BaseM (ConwayUTXOW era) ~ ShelleyBase
+  , PredicateFailure (ConwayUTXOW era) ~ BabbageUtxowPredFailure era
+  , Event (ConwayUTXOW era) ~ AlonzoUtxowEvent era
+  ) =>
+  Embed (BabbageUTXO era) (ConwayUTXOW era)
+  where
+  wrapFailed = UtxoFailure
+  wrapEvent = WrappedShelleyEraEvent . UtxoEvent