+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Version history for `cardano-ledger-core`
 
-## 1.3.0.1
+## 1.3.1.0
 
-*
+* Addition of `getPoolCertTxCert`
 
 ## 1.3.0.0